requirements.txt
python-telegram-bot==12.4.2
flask
gunicorn
requests
selenium
pandas
configparser